backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
Pillow
openai
nltk
textblob
pyttsx3
python-dotenv
diffusers 
transformers 
torch
requests
elevenlabs